HU-06
Fix(web): Add new component Gallery
