+++ Frontend/turismo_front/src/Components/gallery/Gallery.module.css
@@ -1,0 +1,281 @@
+.gallery {
+    display: flex;
+    justify-content: center;
+    gap: 10px;
+}
+
+.mainImage img {
+    width: 30vw;
+    height: 60vh;
+    object-fit: cover;
+    border-radius: 10px;
+    border-top-right-radius: 0;
+    border-bottom-right-radius: 0;
+    cursor: pointer;
+    transition: filter 0.3s ease-in-out;
+}
+
+.mainImage img:hover {
+    filter: brightness(0.8);
+}
+
+
+.sideImages {
+    display: flex;
+    flex-wrap: wrap;
+    width: 40%;
+    gap: 10px;
+    position: relative;
+}
+
+.sideImages img {
+    width: 48%;
+    height: 29vh;
+    object-fit: cover;
+    border-radius: 10px;
+    cursor: pointer;
+    transition: filter 0.3s ease-in-out;
+}
+
+.sideImages img:hover {
+    filter: brightness(0.8);
+}
+
+.sideImages img:first-child {
+    border-radius: 0;
+}
+
+.sideImages img:nth-child(3){
+    border-radius: 0;
+}
+
+.sideImages img:nth-child(2) {
+    border-bottom-left-radius: 0; 
+    border-top-left-radius: 0;
+    border-bottom-right-radius: 0;
+}
+
+.sideImages img:nth-child(4) {
+    border-bottom-left-radius: 0; 
+    border-top-left-radius: 0;
+    border-top-right-radius: 0;
+}
+
+.viewMoreBtn {
+    position: absolute;
+    bottom: 20px;
+    right: 30px;
+    background: rgba(0, 0, 0, 0.7);
+    color: white;
+    padding: 8px;
+    border: none;
+    border-radius: 5px;
+    cursor: pointer;
+}
+
+.modal {
+    position: fixed;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    background: rgba(0, 0, 0, 0.8);
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    animation: fadeInModal 0.3s ease-out forwards;
+}
+
+.modalContent {
+    background: white;
+    padding: 20px;
+    border-radius: 10px;
+    width: 70%;
+    height: 80vh;
+    margin-top: 8vh;
+    text-align: center;
+    overflow-y: auto; 
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    position: relative;
+}
+
+.secondModal {
+    position: fixed;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    background: rgba(0, 0, 0, 0.8);
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    animation: fadeInModal 0.3s ease-out forwards;
+}
+
+.secondModalContent {
+    background: black;
+    padding: 20px;
+    border-radius: 10px;
+    width: 70%;
+    height: 80vh;
+    margin-top: 8vh;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center; 
+    position: relative;
+    text-align: center;
+}
+
+.modalGallery {
+    display: flex;
+    flex-wrap: wrap;
+    gap: 10px;
+    justify-content: center;
+    max-height: 70vh; 
+    overflow-y: auto; 
+    padding: 10px;
+    margin-top: 15px;
+}
+
+
+.modalGallery img {
+    max-width: 100%;
+    height: auto;
+    max-height: 25vh;
+    object-fit: cover;
+    transition: transform 0.3s ease-in-out;
+    cursor: pointer;
+    transition: filter 0.2s ease-in-out;
+}
+
+.modalGallery img:hover {
+    filter: brightness(0.9);
+}
+
+.closeBtn {
+    position: absolute;
+    top: 10px;
+    right: 15px;
+    background: red;
+    color: white;
+    border: none;
+    padding: 5px 10px;
+    font-size: .8rem;
+    cursor: pointer;
+    border-radius: 10%;
+    transition: background 0.3s ease-in-out, transform 0.2s ease-in-out;
+}
+
+
+.closeBtn:hover {
+    background: darkred;
+    transform: scale(1.1);
+}
+
+.imageControls {
+    margin-top: 10px;
+    display: flex;
+    gap: 10px;
+}
+
+.imageControls button {
+    background: #8ecae6;
+    color: #023047;
+    border: none;
+    padding: 8px 12px;
+    font-size: 1rem;
+    cursor: pointer;
+    border-radius: 5px;
+    transition: background 0.3s ease-in-out;
+}
+
+.imageControls button:hover {
+    background: #555;
+}
+
+.largeImage {
+    max-width: 100%;
+    max-height: 70vh;
+    object-fit: contain; 
+}
+
+@media screen and (max-width: 1450px) {
+    .mainImage img {
+        width: 40vw;
+        height: 50vh;
+    }
+
+    .sideImages {
+        width: 50%;
+    }
+
+    .sideImages img {
+        height: 24vh;
+    }
+}
+
+@media screen and (max-width: 1200px) {
+    .mainImage img {
+        width: 40vw;
+        height: 50vh;
+    }
+
+    .sideImages {
+        width: 50%;
+    }
+
+    .sideImages img {
+        height: 24vh;
+    }
+}
+
+@media screen and (max-width: 992px) {
+    .mainImage img {
+        width: 35vw;
+        height: 45vh;
+    }
+
+    .sideImages {
+        width: 45%;
+    }
+
+    .sideImages img {
+        height: 22vh;
+    }
+}
+
+@media screen and (max-width: 768px) {
+    .mainImage img {
+        width: 45vw;
+        height: 35vh;
+    }
+
+    .sideImages {
+        width: 50%;
+    }
+
+    .sideImages img {
+        height: 17vh;
+    }
+}
+
+@media screen and (max-width: 730px) {
+    .gallery {
+        display: none;
+    }
+}
+
+@keyframes fadeInModal {
+    from {
+        opacity: 0;
+        transform: translateY(-20px);
+    }
+    to {
+        opacity: 1;
+        transform: translateY(0);
+    }
+}
+
