+++ Frontend/turismo_front/src/Components/pages/package/Package.jsx
@@ -3,7 +3,7 @@ import { getOnepackage } from '../../services/getOnePackage';
 import { data } from 'react-router-dom';
 import styles from "./Package.module.css";
 import { Link } from 'react-router-dom';
-import ControlledCarousel from "../../carousel/Carousel";
+import Gallery from '../../gallery/Gallery';
 
 const Package = () => {
     /*const { id } = useParams();
@@ -57,35 +57,9 @@ const Package = () => {
             },
         ],
     };
-    /////////////////
-    const [isMobile, setIsMobile] = useState(window.innerWidth <= 730);
-
-    useEffect(() => {
-        const handleResize = () => setIsMobile(window.innerWidth <= 730);
-        window.addEventListener("resize", handleResize);
-        return () => window.removeEventListener("resize", handleResize);
-    }, []);
-    //////////////////
     const { title, description, categories } = travelPackage;
     const [{ title: categoryTitle, description: categoryDescription, price, currency, restrictions, mediaCategories }] = categories || [];
     const [{ mediaTitle, mediaDescription, media }] = mediaCategories || [];
-
-    const [isModalOpen, setIsModalOpen] = useState(false);
-    const [isSecondModalOpen, setIsSecondModalOpen] = useState(false);
-    const [selectedImageIndex, setSelectedImageIndex] = useState(0);
-
-    const openSecondModal = (index) => {
-        setSelectedImageIndex(index);
-        setIsSecondModalOpen(true);
-    };
-
-    const changeImage = (direction) => {
-        if (direction === "next") {
-            setSelectedImageIndex((prevIndex) => (prevIndex + 1) % images.length);
-        } else {
-            setSelectedImageIndex((prevIndex) => (prevIndex - 1 + images.length) % images.length);
-        }
-    };
 
     return (
         <div className={styles.mainContainer}>
@@ -93,53 +67,7 @@ const Package = () => {
             <div className={styles.titleContainer}>
                 <h1>{title}</h1>
             </div>
-            <div className={styles.packageContainer}>
-                {isMobile ? (
-                    <ControlledCarousel images={images.map(src => ({ src }))} openModal={() => setIsModalOpen(true)} isModalOpen={isModalOpen} />
-                ) : (
-                    <div className={styles.gallery}>
-                        <div className={styles.mainImage}>
-                            <img src={images[0]} alt="Main" onClick={() => setIsModalOpen(true)} />
-                        </div>
-                        <div className={styles.sideImages}>
-                            {images.slice(1, 5).map((img, index) => (
-                                <img key={index} src={img} alt={`Gallery ${index + 1}`} onClick={() => setIsModalOpen(true)} />
-                            ))}
-                            <button className={styles.viewMoreBtn} onClick={() => setIsModalOpen(true)}>
-                                Ver más fotos
-                            </button>
-                        </div>
-                    </div>
-                )}
-
-            {/* Modal de imágenes */}
-            {isModalOpen && (
-                <div className={styles.modal}>
-                    <div className={styles.modalContent}>
-                        <button className={styles.closeBtn} onClick={() => setIsModalOpen(false)}><i class="bi bi-x"></i></button>
-                        <div className={styles.modalGallery}>
-                            {images.map((img, index) => (
-                                <img key={index} src={img} alt={`Modal ${index + 1}`} onClick={() => openSecondModal(index)} />
-                            ))}
-                        </div>
-                    </div>
-                </div>
-            )}
-
-            {/* Segundo modal para ver la imagen en grande */}
-            {isSecondModalOpen && (
-                <div className={styles.secondModal}>
-                    <div className={styles.secondModalContent}>
-                        <button className={styles.closeBtn} onClick={() => setIsSecondModalOpen(false)}><i class="bi bi-x"></i></button>
-                        <img src={images[selectedImageIndex]} alt="Selected" className={styles.largeImage} />
-                        <div className={styles.imageControls}>
-                            <button onClick={() => changeImage("prev")}><i class="bi bi-arrow-left-short"></i></button>
-                            <button onClick={() => changeImage("next")}><i class="bi bi-arrow-right-short"></i></button>
-                        </div>
-                    </div>
-                </div>
-            )}
-        </div>
+            <Gallery images={images} />
         </div >
     );
 }
